Refactor set_titles to a title.py module.
